Connect to server specified in serverless bundle
Before this change, the driver, when connecting to a serverless cluster,
would try to resolve host-id.* domains as a proxy IP for each node.

This commit aligns the behavior with scylladb/python-driver#177 and
scylladb/gocql#106. Now, the "Server" field of bundle is the single
hostname/IP the driver will use for proxy connection.

This simplifies the setup we had in integration tests, removing the
necessity of hacks to work around previous behavior (driver trying to
resolve host-id.* domains).

Fixes #164
Fixes #167
Fixes #169