Wait for agreement in test in SchemaAgreementTest
This is an alternative solution to d962bf603682ed0e8b8defc01f2a99643b79320b.

That commit tried to deal with the fact, that 
should_unset_flag_on_failed_agreement would leave the schema in 
disagreement. If another test ran after that one, it would start with
schema in disagreement. 

The previous solution was to set priority=1, so that this test would
run after other tests. The problem with that solution is that all other
tests from SchemaAgreementTests would run at one moment, then some 
unrelated tests would run and finally should_unset_flag_on_failed_agreement
would run. Those unrelated tests could screw up the execution of
should_unset_flag_on_failed_agreement, as seen in #170.

Fix that problem by running a dummy query with long schema agreement
timeout, so that after this test runs, the schema is in agreement and
the state of affairs is clean.

Fixes #170