Pin ubuntu runner images to 20.04 version
Instead of running GitHub Actions on ubuntu-latest (which recently
changed to 22.04) pin a specific version of 20.04.

When I was first making the workflows I used ubuntu-latest as a way
to have motivation to have the workflow code modern. With more
experience I see that it unnecessarily adds work maintaining the repo.