Add RackAwareRoundRobinPolicy
RackAwareRoundRobinPolicy is a modified version of 
DCAwareRoundRobinPolicy. The new policy shares much of the code with
the original policy, but changes the logic of newQueryPlan - it will
return hosts in local rack first, followed by remote racks in local DC, 
followed by remote DC.

In contrast to the approach in #199, this policy does not change the 
distance() method - it will return LOCAL for both local rack in local DC
and remote racks in local DC. This is a deliberate choice, as many other
places in the driver look at the distance and introducing a new distance
type (e.g. LOCAL_REMOTERACK) would necessitate many changes. One example
of that is PoolingOptions, where the user can specify how many 
connections should be made to LOCAL/REMOTE nodes - adding another type
of distance would force the user to remember to change yet another
configuration parameter.

Integration tests in RackAwareRoundRobinPolicyTest are based on tests
in DCAwareRoundRobinPolicyTest, as much of the behavior should be
the same between the policies. A couple new tests were introduced to
specifically test the rack-aware feature in local DC.